Add a sphere shape

+++ src/float.rs
@@ -1,2 +1,2 @@
 
-pub type Float = f32;
+pub type Float = f64;
